BAU: Update linting config
Add the prettier plugin to ESlint.
The plugin was already installed in package.json but it wasn't used, as it needs to be explicitly added to ESlint config.

ESlint now highlights prettier rules in the IDE and fixes the errors if the `--fix` flag is used.

Exclude package-lock.json from prettier.
Update linting scripts.